🐛 修复`gs角色排名龙王`, 如果排名靠前找不到自己

+++ GenshinUID/genshinuid_enka/draw_role_rank.py
@@ -58,8 +58,8 @@ async def draw_role_rank_img(
 
         fit = rank_data[char_id]['calculations']['fit']
         calculation_id = fit['calculationId']
-        r0 = int(fit["result"])
-        r1 = int(r0 + 0.01)
+        r0 = float(fit["result"])
+        r1 = r0 + 0.01
         raw_data = await cv_api.get_sort_list(char_id, calculation_id, r1)
         tag = '角色附近'
     else:
